Extract io_utils, layer_split_types, and draft_ipc from test_dflash.cpp
- src/common/io_utils.h: I/O helpers (read/write int32 files, fd ops, argmax)
- src/qwen35/layer_split_types.h: TargetLayerSplitShard, ActivationPair,
  LayerSplitRuntimeConfig, find_target_shard()
- src/qwen35/draft_ipc.{h,cpp}: DFlashDraftIpcClient class,
  run_dflash_draft_ipc_daemon(), copy_capture_slice_to_remote_draft()

test_dflash.cpp: 4498 -> 3937 lines (-561)
