+++ dflash/CMakeLists.txt
@@ -226,6 +226,7 @@ add_library(dflash27b STATIC
     src/qwen35/draft_feature_mirror.cpp
     src/qwen35/graph_builders.cpp
     src/qwen35/feature_copy.cpp
+    src/qwen35/draft_ipc.cpp
     src/qwen35/qwen35_backend.cpp
     src/qwen35/qwen35_daemon.cpp
     src/common/sampler.cpp
