+++ dflash/test/test_dflash.cpp
@@ -107,46 +107,19 @@ static std::mt19937_64 g_sampler_rng{std::random_device{}()};
     ( ((w).eos_chat_id >= 0 && (tok) == (w).eos_chat_id)                  \
    || ((w).eos_id      >= 0 && (tok) == (w).eos_id     ) )
 
-// ─── Small utilities ──────────────────────────────────────────────
-
-static std::vector<int32_t> read_int32_file(const std::string & path) {
-    std::ifstream f(path, std::ios::binary | std::ios::ate);
-    if (!f) return {};
-    auto sz = (size_t)f.tellg();
-    f.seekg(0);
-    std::vector<int32_t> out(sz / sizeof(int32_t));
-    f.read((char *)out.data(), sz);
-    return out;
-}
-
-static bool write_int32_file(const std::string & path, const std::vector<int32_t> & v) {
-    std::ofstream f(path, std::ios::binary);
-    if (!f) return false;
-    f.write((const char *)v.data(), v.size() * sizeof(int32_t));
-    return (bool)f;
-}
-
-static void stream_emit_fd(int stream_fd, int32_t tok) {
-    if (stream_fd < 0) return;
-#if defined(_WIN32)
-    DWORD written = 0;
-    const int32_t v = tok;
-    WriteFile((HANDLE)(intptr_t)stream_fd, &v, sizeof(v), &written, nullptr);
-#else
-    const int32_t v = tok;
-    const ssize_t n = ::write(stream_fd, &v, sizeof(v));
-    if (n < 0) {
-        // Best-effort stream path; daemon loop handles downstream EOF/termination.
-    }
+// ─── Small utilities — extracted to src/common/io_utils.h ──────────
+#include "io_utils.h"
+using dflash27b::read_int32_file;
+using dflash27b::write_int32_file;
+using dflash27b::stream_emit_fd;
+using dflash27b::argmax_f32;
+using dflash27b::write_binary_file;
+using dflash27b::read_binary_file_exact;
+using dflash27b::read_line_tail;
+#if !defined(_WIN32)
+using dflash27b::read_exact_fd;
+using dflash27b::write_exact_fd;
 #endif
-}
-
-static int argmax_f32(const float * x, int n) {
-    int best = 0;
-    float bv = x[0];
-    for (int i = 1; i < n; i++) if (x[i] > bv) { bv = x[i]; best = i; }
-    return best;
-}
 
 // CPU sampler chain (SamplerCfg / sample_logits / parse_sampler_token) lives
 // in src/sampler.{h,cpp} and is shared with src/laguna_daemon.cpp. Behaviour
@@ -211,59 +184,14 @@ using dflash27b::build_target_step_tree;
 using dflash27b::build_draft_step;
 using dflash27b::build_lm_head_projection_step;
 
-struct TargetLayerSplitShard {
-    int gpu = 0;
-    int layer_begin = 0;
-    int layer_end = 0;
-    ggml_backend_t backend = nullptr;
-    TargetWeights weights;
-    TargetCache cache;
-    StepGraph layer_graph;
-};
-
-struct ActivationPair {
-    ggml_context * ctx = nullptr;
-    ggml_backend_buffer_t buf = nullptr;
-    ggml_tensor * a = nullptr;
-    ggml_tensor * b = nullptr;
-    ggml_backend_t backend = nullptr;
-    int n_tokens = 0;
-};
-
-static void activation_pair_free(ActivationPair & p) {
-    if (p.buf) { ggml_backend_buffer_free(p.buf); p.buf = nullptr; }
-    if (p.ctx) { ggml_free(p.ctx); p.ctx = nullptr; }
-    p.a = nullptr;
-    p.b = nullptr;
-    p.backend = nullptr;
-    p.n_tokens = 0;
-}
-
-static bool activation_pair_init(ActivationPair & p,
-                                 ggml_backend_t backend,
-                                 int hidden,
-                                 int n_tokens) {
-    activation_pair_free(p);
-    if (n_tokens <= 0) return false;
-    p.backend = backend;
-    p.n_tokens = n_tokens;
-    ggml_init_params ip{};
-    ip.mem_size = (size_t)8 * ggml_tensor_overhead() + 16 * 1024;
-    ip.mem_buffer = nullptr;
-    ip.no_alloc = true;
-    p.ctx = ggml_init(ip);
-    if (!p.ctx) return false;
-    p.a = ggml_new_tensor_2d(p.ctx, GGML_TYPE_F32, hidden, n_tokens);
-    p.b = ggml_new_tensor_2d(p.ctx, GGML_TYPE_F32, hidden, n_tokens);
-    ggml_set_name(p.a, "target_split_act_a");
-    ggml_set_name(p.b, "target_split_act_b");
-    p.buf = ggml_backend_alloc_ctx_tensors(p.ctx, backend);
-    if (!p.buf) {
-        activation_pair_free(p);
-        return false;
-    }
-    return true;
-}
+// ─── Layer split types — extracted to src/qwen35/layer_split_types.h ──
+#include "layer_split_types.h"
+using dflash27b::LayerSplitRuntimeConfig;
+using dflash27b::TargetLayerSplitShard;
+using dflash27b::ActivationPair;
+using dflash27b::activation_pair_free;
+using dflash27b::activation_pair_init;
+using dflash27b::find_target_shard;
 
 static bool parse_int_list(const char * text, std::vector<int> & out) {
     out.clear();
@@ -297,273 +225,12 @@ static bool parse_float_list(const char * text, std::vector<double> & out) {
     return !out.empty();
 }
 
-static bool write_binary_file(const std::string & path, const void * data, size_t bytes) {
-    std::ofstream f(path, std::ios::binary);
-    if (!f) return false;
-    if (bytes > 0) f.write((const char *)data, (std::streamsize)bytes);
-    return (bool)f;
-}
-
-static bool read_binary_file_exact(const std::string & path, void * data, size_t bytes) {
-    std::ifstream f(path, std::ios::binary);
-    if (!f) return false;
-    if (bytes > 0) f.read((char *)data, (std::streamsize)bytes);
-    return (bool)f;
-}
-
-static std::string read_line_tail(std::istringstream & iss) {
-    std::string tail;
-    std::getline(iss, tail);
-    const size_t first = tail.find_first_not_of(" \t");
-    if (first == std::string::npos) return {};
-    if (first > 0) tail.erase(0, first);
-    return tail;
-}
-
-#if !defined(_WIN32)
-static bool read_exact_fd(int fd, void * data, size_t bytes) {
-    char * p = (char *)data;
-    size_t done = 0;
-    while (done < bytes) {
-        ssize_t n = ::read(fd, p + done, bytes - done);
-        if (n == 0) return false;
-        if (n < 0) {
-            if (errno == EINTR) continue;
-            return false;
-        }
-        done += (size_t)n;
-    }
-    return true;
-}
-
-static bool write_exact_fd(int fd, const void * data, size_t bytes) {
-    const char * p = (const char *)data;
-    size_t done = 0;
-    while (done < bytes) {
-        ssize_t n = ::write(fd, p + done, bytes - done);
-        if (n < 0) {
-            if (errno == EINTR) continue;
-            return false;
-        }
-        done += (size_t)n;
-    }
-    return true;
-}
-#endif
-
-class DFlashDraftIpcClient {
-public:
-    DFlashDraftIpcClient() = default;
-    DFlashDraftIpcClient(const DFlashDraftIpcClient &) = delete;
-    DFlashDraftIpcClient & operator=(const DFlashDraftIpcClient &) = delete;
-    ~DFlashDraftIpcClient() { close(); }
-
-    bool start(const std::string & bin,
-               const std::string & draft_path,
-               int draft_gpu,
-               int ring_cap,
-               const std::string & work_dir) {
-#if defined(_WIN32)
-        (void)bin; (void)draft_path; (void)draft_gpu; (void)ring_cap; (void)work_dir;
-        std::fprintf(stderr, "DFlash draft IPC is only implemented on POSIX hosts\n");
-        return false;
-#else
-        close();
-        if (bin.empty() || draft_path.empty() || ring_cap <= 0) return false;
-        if (!init_work_dir(work_dir)) return false;
-
-        int cmd_pipe[2] = {-1, -1};
-        int stream_pipe[2] = {-1, -1};
-        if (::pipe(cmd_pipe) != 0 || ::pipe(stream_pipe) != 0) {
-            std::fprintf(stderr, "draft-ipc pipe failed: %s\n", std::strerror(errno));
-            if (cmd_pipe[0] >= 0) ::close(cmd_pipe[0]);
-            if (cmd_pipe[1] >= 0) ::close(cmd_pipe[1]);
-            if (stream_pipe[0] >= 0) ::close(stream_pipe[0]);
-            if (stream_pipe[1] >= 0) ::close(stream_pipe[1]);
-            return false;
-        }
-
-        pid_ = ::fork();
-        if (pid_ < 0) {
-            std::fprintf(stderr, "draft-ipc fork failed: %s\n", std::strerror(errno));
-            ::close(cmd_pipe[0]); ::close(cmd_pipe[1]);
-            ::close(stream_pipe[0]); ::close(stream_pipe[1]);
-            pid_ = -1;
-            return false;
-        }
-        if (pid_ == 0) {
-            ::dup2(cmd_pipe[0], STDIN_FILENO);
-            ::close(cmd_pipe[0]);
-            ::close(cmd_pipe[1]);
-            ::close(stream_pipe[0]);
-
-            const std::string cap_arg = "--ring-cap=" + std::to_string(ring_cap);
-            const std::string gpu_arg = "--draft-gpu=" + std::to_string(std::max(0, draft_gpu));
-            const std::string fd_arg = "--stream-fd=" + std::to_string(stream_pipe[1]);
-            ::execl(bin.c_str(), bin.c_str(),
-                    "--draft-ipc-daemon", draft_path.c_str(),
-                    cap_arg.c_str(), gpu_arg.c_str(), fd_arg.c_str(),
-                    (char *)nullptr);
-            std::fprintf(stderr, "draft-ipc exec failed: %s: %s\n",
-                         bin.c_str(), std::strerror(errno));
-            _exit(127);
-        }
-
-        ::close(cmd_pipe[0]);
-        ::close(stream_pipe[1]);
-        stream_fd_ = stream_pipe[0];
-        cmd_ = ::fdopen(cmd_pipe[1], "w");
-        if (!cmd_) {
-            std::fprintf(stderr, "draft-ipc fdopen failed: %s\n", std::strerror(errno));
-            ::close(cmd_pipe[1]);
-            close();
-            return false;
-        }
-        int32_t status = -1;
-        if (!read_exact_fd(stream_fd_, &status, sizeof(status)) || status != 0) {
-            std::fprintf(stderr, "draft-ipc daemon did not become ready (status=%d)\n", status);
-            close();
-            return false;
-        }
-        ring_cap_ = ring_cap;
-        active_ = true;
-        std::printf("[draft-ipc] ready bin=%s gpu=%d ring_cap=%d work_dir=%s\n",
-                    bin.c_str(), draft_gpu, ring_cap, work_dir_.c_str());
-        return true;
-#endif
-    }
-
-    bool send_feature_slice(int capture_idx,
-                            int start_pos,
-                            int n_tokens,
-                            const std::vector<float> & slice) {
-#if defined(_WIN32)
-        (void)capture_idx; (void)start_pos; (void)n_tokens; (void)slice;
-        return false;
-#else
-        if (!active_ || !cmd_ || n_tokens <= 0) return false;
-        const size_t expected = (size_t)n_tokens * DFLASH27B_TARGET_HIDDEN;
-        if (slice.size() != expected) return false;
-        const std::string path = next_path("feature");
-        if (!write_binary_file(path, slice.data(), slice.size() * sizeof(float))) {
-            std::fprintf(stderr, "draft-ipc write feature failed: %s\n", path.c_str());
-            return false;
-        }
-        std::fprintf(cmd_, "feature_slice %d %d %d %s\n",
-                     capture_idx, start_pos, n_tokens, path.c_str());
-        std::fflush(cmd_);
-        int32_t status = -1;
-        const bool ok = read_exact_fd(stream_fd_, &status, sizeof(status)) && status == 0;
-        std::remove(path.c_str());
-        if (!ok) {
-            std::fprintf(stderr, "draft-ipc feature_slice failed status=%d\n", status);
-        }
-        return ok;
-#endif
-    }
-
-    bool propose(int committed,
-                 int ctx_len,
-                 const std::vector<float> & noise_embed,
-                 std::vector<float> & hidden_out) {
-#if defined(_WIN32)
-        (void)committed; (void)ctx_len; (void)noise_embed; (void)hidden_out;
-        return false;
-#else
-        if (!active_ || !cmd_ || ctx_len <= 0) return false;
-        const size_t noise_expected =
-            (size_t)DFLASH27B_TARGET_HIDDEN * DFLASH27B_DRAFT_BLOCK_SIZE;
-        if (noise_embed.size() != noise_expected) return false;
-        const std::string path = next_path("noise");
-        if (!write_binary_file(path, noise_embed.data(), noise_embed.size() * sizeof(float))) {
-            std::fprintf(stderr, "draft-ipc write noise failed: %s\n", path.c_str());
-            return false;
-        }
-        std::fprintf(cmd_, "propose %d %d %s\n", committed, ctx_len, path.c_str());
-        std::fflush(cmd_);
-        int32_t status = -1;
-        bool ok = read_exact_fd(stream_fd_, &status, sizeof(status)) && status == 0;
-        if (ok) {
-            hidden_out.assign(noise_expected, 0.0f);
-            ok = read_exact_fd(stream_fd_, hidden_out.data(),
-                               hidden_out.size() * sizeof(float));
-        }
-        std::remove(path.c_str());
-        if (!ok) {
-            std::fprintf(stderr, "draft-ipc propose failed status=%d\n", status);
-        }
-        return ok;
-#endif
-    }
-
-    bool active() const { return active_; }
-    int ring_cap() const { return ring_cap_; }
-
-    void close() {
-#if !defined(_WIN32)
-        if (cmd_) {
-            std::fclose(cmd_);
-            cmd_ = nullptr;
-        }
-        if (stream_fd_ >= 0) {
-            ::close(stream_fd_);
-            stream_fd_ = -1;
-        }
-        if (pid_ > 0) {
-            int status = 0;
-            ::waitpid(pid_, &status, 0);
-            pid_ = -1;
-        }
-        if (owns_work_dir_ && !work_dir_.empty()) {
-            ::rmdir(work_dir_.c_str());
-        }
-#endif
-        active_ = false;
-        ring_cap_ = 0;
-    }
-
-private:
-#if !defined(_WIN32)
-    bool init_work_dir(const std::string & requested) {
-        if (!requested.empty()) {
-            work_dir_ = requested;
-            owns_work_dir_ = false;
-            if (::mkdir(work_dir_.c_str(), 0700) != 0 && errno != EEXIST) {
-                std::fprintf(stderr, "draft-ipc mkdir failed: %s: %s\n",
-                             work_dir_.c_str(), std::strerror(errno));
-                return false;
-            }
-            return true;
-        }
-        const char * tmp = std::getenv("TMPDIR");
-        std::string templ = std::string(tmp && *tmp ? tmp : "/tmp") +
-                            "/dflash-draft-ipc-XXXXXX";
-        std::vector<char> buf(templ.begin(), templ.end());
-        buf.push_back('\0');
-        char * dir = ::mkdtemp(buf.data());
-        if (!dir) {
-            std::fprintf(stderr, "draft-ipc mkdtemp failed: %s\n", std::strerror(errno));
-            return false;
-        }
-        work_dir_ = dir;
-        owns_work_dir_ = true;
-        return true;
-    }
-
-    std::string next_path(const char * prefix) {
-        return work_dir_ + "/" + prefix + "_" + std::to_string(seq_++) + ".bin";
-    }
-
-    pid_t pid_ = -1;
-    FILE * cmd_ = nullptr;
-    int stream_fd_ = -1;
-    std::string work_dir_;
-    int seq_ = 0;
-    bool owns_work_dir_ = false;
-#endif
-    bool active_ = false;
-    int ring_cap_ = 0;
-};
+// ─── Draft IPC — extracted to src/qwen35/draft_ipc.{h,cpp} ──
+#include "draft_ipc.h"
+using dflash27b::DFlashDraftIpcClient;
+using dflash27b::copy_capture_slice_to_remote_draft;
+using dflash27b::stream_status;
+using dflash27b::run_dflash_draft_ipc_daemon;
 
 // ─── GGUF inspection — extracted to src/common/gguf_inspect.{h,cpp} ──
 #include "gguf_inspect.h"
@@ -576,239 +243,11 @@ static int inspect_target_layer_count(const char * target_path) {
 #include "layer_split_utils.h"
 using dflash27b::compute_layer_ranges;
 
-static TargetLayerSplitShard * find_target_shard(
-        std::vector<TargetLayerSplitShard> & shards,
-        int layer_idx) {
-    for (auto & shard : shards) {
-        if (layer_idx >= shard.layer_begin && layer_idx < shard.layer_end) {
-            return &shard;
-        }
-    }
-    return nullptr;
-}
-
 // ─── Feature copy helpers — extracted to src/qwen35/feature_copy.{h,cpp} ──
 #include "feature_copy.h"
 using dflash27b::target_capture_index;
 using dflash27b::copy_capture_slice_to_draft_ring;
 using dflash27b::copy_feature_ring_range_to_tensor;
-
-static bool copy_capture_slice_to_remote_draft(
-        DFlashDraftIpcClient & remote,
-        int capture_idx,
-        const ggml_tensor * act_out,
-        ggml_backend_t src_backend,
-        int chunk_start,
-        int start_pos,
-        int n_tokens) {
-    if (!remote.active() || !act_out || capture_idx < 0 || n_tokens <= 0) return true;
-    const int hidden = DFLASH27B_TARGET_HIDDEN;
-    const size_t row_bytes = (size_t)hidden * sizeof(float);
-    const size_t src_stride = act_out->nb[1];
-    std::vector<float> host((size_t)n_tokens * hidden);
-    ggml_backend_synchronize(src_backend);
-    if (src_stride == row_bytes) {
-        ggml_backend_tensor_get(act_out, host.data(),
-                                (size_t)chunk_start * src_stride,
-                                row_bytes * (size_t)n_tokens);
-    } else {
-        for (int i = 0; i < n_tokens; i++) {
-            ggml_backend_tensor_get(act_out,
-                                    host.data() + (size_t)i * hidden,
-                                    (size_t)(chunk_start + i) * src_stride,
-                                    row_bytes);
-        }
-    }
-    return remote.send_feature_slice(capture_idx, start_pos, n_tokens, host);
-}
-
-static bool stream_status(int stream_fd, int32_t status) {
-#if defined(_WIN32)
-    (void)stream_fd; (void)status;
-    return false;
-#else
-    return write_exact_fd(stream_fd, &status, sizeof(status));
-#endif
-}
-
-static int run_dflash_draft_ipc_daemon(const char * draft_path,
-                                       int ring_cap,
-                                       int draft_gpu,
-                                       int stream_fd) {
-#if defined(_WIN32)
-    (void)draft_path; (void)ring_cap; (void)draft_gpu; (void)stream_fd;
-    std::fprintf(stderr, "DFlash draft IPC daemon is only implemented on POSIX hosts\n");
-    return 2;
-#else
-    if (!draft_path || ring_cap <= 0 || stream_fd < 0) {
-        std::fprintf(stderr, "usage: test_dflash --draft-ipc-daemon <draft> --ring-cap=N --stream-fd=FD [--draft-gpu=N]\n");
-        return 2;
-    }
-
-    ggml_backend_t backend = ggml_backend_cuda_init(std::max(0, draft_gpu));
-    if (!backend) {
-        std::fprintf(stderr, "[draft-ipc-daemon] backend init failed gpu=%d\n", draft_gpu);
-        stream_status(stream_fd, -1);
-        return 1;
-    }
-
-    DraftWeights draft_weights;
-    std::string dp(draft_path);
-    bool draft_ok = false;
-    if (dp.size() >= 5 && dp.substr(dp.size() - 5) == ".gguf") {
-        draft_ok = load_draft_gguf(draft_path, backend, draft_weights);
-    } else {
-        draft_ok = load_draft_safetensors(draft_path, backend, draft_weights);
-    }
-    if (!draft_ok) {
-        std::fprintf(stderr, "[draft-ipc-daemon] draft load failed: %s\n",
-                     dflash27b_last_error());
-        stream_status(stream_fd, -1);
-        ggml_backend_free(backend);
-        return 1;
-    }
-
-    DraftFeatureMirror feature_ring;
-    if (!draft_feature_mirror_init(feature_ring, backend, draft_gpu, draft_gpu, ring_cap)) {
-        std::fprintf(stderr, "[draft-ipc-daemon] feature ring init failed cap=%d gpu=%d\n",
-                     ring_cap, draft_gpu);
-        stream_status(stream_fd, -1);
-        free_draft_weights(draft_weights);
-        ggml_backend_free(backend);
-        return 1;
-    }
-
-    std::fprintf(stderr, "[draft-ipc-daemon] ready gpu=%d ring_cap=%d\n",
-                 draft_gpu, ring_cap);
-    stream_status(stream_fd, 0);
-
-    const int hidden = DFLASH27B_TARGET_HIDDEN;
-    const int q_len = DFLASH27B_DRAFT_BLOCK_SIZE;
-    StepGraph draft_sg;
-    std::vector<float> noise_embed((size_t)hidden * q_len);
-    std::vector<int32_t> pos_q(q_len);
-    std::vector<int32_t> pos_k;
-    std::vector<float> hidden_out((size_t)hidden * q_len);
-
-    std::string line;
-    while (std::getline(std::cin, line)) {
-        std::istringstream iss(line);
-        std::string cmd;
-        iss >> cmd;
-        if (cmd == "quit" || cmd == "exit") {
-            break;
-        }
-        if (cmd == "feature_slice") {
-            int capture_idx = -1;
-            int start_pos = -1;
-            int n_tokens = 0;
-            iss >> capture_idx >> start_pos >> n_tokens;
-            std::string path = read_line_tail(iss);
-            if (capture_idx < 0 || capture_idx >= DFLASH27B_DRAFT_N_TARGET_LAYERS ||
-                start_pos < 0 || n_tokens <= 0 || path.empty()) {
-                std::fprintf(stderr, "[draft-ipc-daemon] bad feature_slice: %s\n",
-                             line.c_str());
-                stream_status(stream_fd, -1);
-                continue;
-            }
-            std::vector<float> slice((size_t)n_tokens * hidden);
-            if (!read_binary_file_exact(path, slice.data(), slice.size() * sizeof(float))) {
-                std::fprintf(stderr, "[draft-ipc-daemon] read feature_slice failed: %s\n",
-                             path.c_str());
-                stream_status(stream_fd, -1);
-                continue;
-            }
-            const size_t dst_stride = feature_ring.target_feat->nb[1];
-            const size_t slice_offset =
-                (size_t)capture_idx * (size_t)hidden * sizeof(float);
-            for (int i = 0; i < n_tokens; i++) {
-                const int slot = (start_pos + i) % feature_ring.cap;
-                const size_t dst_off = (size_t)slot * dst_stride + slice_offset;
-                ggml_backend_tensor_set(feature_ring.target_feat,
-                                        slice.data() + (size_t)i * hidden,
-                                        dst_off,
-                                        (size_t)hidden * sizeof(float));
-            }
-            ggml_backend_synchronize(backend);
-            stream_status(stream_fd, 0);
-            continue;
-        }
-        if (cmd == "propose") {
-            int committed = -1;
-            int ctx_len = 0;
-            iss >> committed >> ctx_len;
-            std::string path = read_line_tail(iss);
-            if (committed < 0 || ctx_len <= 0 || ctx_len > feature_ring.cap || path.empty()) {
-                std::fprintf(stderr, "[draft-ipc-daemon] bad propose: %s\n",
-                             line.c_str());
-                stream_status(stream_fd, -1);
-                continue;
-            }
-            if (!read_binary_file_exact(path, noise_embed.data(),
-                                        noise_embed.size() * sizeof(float))) {
-                std::fprintf(stderr, "[draft-ipc-daemon] read noise failed: %s\n",
-                             path.c_str());
-                stream_status(stream_fd, -1);
-                continue;
-            }
-
-            int mirror_slot0 = 0;
-            const bool use_mirror_view =
-                draft_feature_mirror_can_view(feature_ring, committed, ctx_len, mirror_slot0);
-            if (!build_draft_step(draft_sg, draft_weights, nullptr, backend,
-                                  ctx_len, use_mirror_view ? &feature_ring : nullptr,
-                                  committed)) {
-                std::fprintf(stderr, "[draft-ipc-daemon] draft build failed\n");
-                stream_status(stream_fd, -1);
-                continue;
-            }
-            if (!use_mirror_view &&
-                !copy_feature_ring_range_to_tensor(feature_ring,
-                                                   draft_sg.target_hidden_cat,
-                                                   committed - ctx_len,
-                                                   ctx_len)) {
-                std::fprintf(stderr, "[draft-ipc-daemon] feature copy failed\n");
-                stream_status(stream_fd, -1);
-                continue;
-            }
-            ggml_backend_tensor_set(draft_sg.inp_embed, noise_embed.data(), 0,
-                                    noise_embed.size() * sizeof(float));
-            pos_k.resize((size_t)ctx_len + q_len);
-            for (int i = 0; i < q_len; i++) pos_q[i] = ctx_len + i;
-            for (int i = 0; i < ctx_len + q_len; i++) pos_k[i] = i;
-            ggml_backend_tensor_set(draft_sg.positions, pos_q.data(), 0,
-                                    pos_q.size() * sizeof(int32_t));
-            ggml_backend_tensor_set(draft_sg.positions_k, pos_k.data(), 0,
-                                    pos_k.size() * sizeof(int32_t));
-            auto st = ggml_backend_graph_compute(backend, draft_sg.gf);
-            if (st != GGML_STATUS_SUCCESS) {
-                std::fprintf(stderr, "[draft-ipc-daemon] draft compute failed status=%d\n",
-                             (int)st);
-                stream_status(stream_fd, -1);
-                continue;
-            }
-            ggml_backend_tensor_get(draft_sg.hidden_states, hidden_out.data(), 0,
-                                    hidden_out.size() * sizeof(float));
-            if (!stream_status(stream_fd, 0) ||
-                !write_exact_fd(stream_fd, hidden_out.data(),
-                                hidden_out.size() * sizeof(float))) {
-                std::fprintf(stderr, "[draft-ipc-daemon] stream write failed\n");
-                break;
-            }
-            continue;
-        }
-        std::fprintf(stderr, "[draft-ipc-daemon] unknown command: %s\n", line.c_str());
-        stream_status(stream_fd, -1);
-    }
-
-    step_graph_destroy(draft_sg);
-    draft_feature_mirror_free(feature_ring);
-    free_draft_weights(draft_weights);
-    ggml_backend_free(backend);
-    std::fprintf(stderr, "[draft-ipc-daemon] stopped\n");
-    return 0;
-#endif
-}
 
 static bool compute_target_split_argmax(
         StepGraph & sg,
